feat(recovery): remove Console Access commands; recovery owns it all
Deletes the 4 legacy Tauri commands that were still registered for
in-flight sessions:
- enable_console_access
- disable_console_access
- console_access_status
- rotate_console_passphrase

Along with their dead-weight types (ConsoleAccessStatus, PolicyGate,
BlobRecord, check_enable_policy, check_rotation_policy, policy_to_err)
and the `confirmed_backup` checkbox-gate code path. Account recovery
is always-on at OAuth signup; there's no opt-in surface left to gate.

Renames `validate_console_passphrase` → `validate_recovery_password`
so the remaining command matches the feature.

Shared plumbing stays in `console_access.rs` (the module name is legacy
debt noted in the docstring — a follow-up PR can rename it or split
into `recovery/http.rs` + `recovery/passphrase.rs`):
- HcfsServerCtx + bearer-auth HTTP helpers (`get_json`,
  `post_json_discard`, `resolve_hcfs_base_url`, `http_err`)
- Passphrase strength scoring (score_passphrase, PassphraseStrength,
  PassphraseVerdict)
- `crypto_to_err` + `short_ss58`

Frontend:
- Deletes `ConsoleAccessSettings.tsx` (already unreferenced after the
  Settings tab removal).
- Deletes `consoleAccess.ts` utility wrapper. The passphrase types and
  the renamed `validateRecoveryPassword` wrapper move into
  `recovery.ts`, so both recovery dialogs import from one place.

Tests retained for the shared scoring + HTTP helpers; the 4 tests for
the deleted policy gates are removed (no policy to test).

Refs: docs/plans/2026-04-14-oauth-account-recovery.md
