+++ src-tauri/src/main.rs
@@ -44,9 +44,7 @@ use crate::billing::queries::{
 use crate::billing::subscriptions::{create_subscription, get_customer_portal_url, get_subscription_data};
 use crate::blockchain::convert::{planck_to_hip_full, to_plancks};
 use crate::blockchain::transfers::compute_max_transferable;
-use crate::console_access::{
-    console_access_status, disable_console_access, enable_console_access, rotate_console_passphrase, validate_console_passphrase,
-};
+use crate::console_access::validate_recovery_password;
 use crate::recovery::{check_recovery_state, mark_recovery_skipped, recover_mnemonic, seal_and_upload_mnemonic};
 use crate::blockchain::queries::{get_account_balance, get_block_timestamp, get_referral_links, get_staking_info, validate_address};
 use crate::blockchain::runtime::{get_wss_endpoint, test_rpc_endpoint_command, update_wss_endpoint_command};
@@ -269,12 +267,8 @@ fn main() {
             planck_to_hip_full,
             compute_max_transferable,
             // Console access
-            console_access_status,
-            validate_console_passphrase,
-            enable_console_access,
-            rotate_console_passphrase,
-            disable_console_access,
             // Account recovery (OAuth-based)
+            validate_recovery_password,
             check_recovery_state,
             recover_mnemonic,
             seal_and_upload_mnemonic,
